Update http_server.py
Great work! I like your use of throwaway variables-- `_`--that tell other programmers that you're deliberately not using those particular returns. And your chunking the files that you're reading is also great: if you were concerned about an infinite file or a file that's too long, then in a real world scenario you might have some kind of breakout condition that exits that loop after your content has grown to a certain size.

Check out a couple of things in my pull request. My _main_ suggestion is using the os.path functions to build your path (starting at line 89). Your logic is excellent (all of the requested file paths should be relative to the web root directory) and the canonical way to build a path in Python would be with os.path.join. The main practical benefit is that os.path.join will build paths that are OS independent: it knows whether to use forward slashes or backwards slashes on the machine that the code is running on. The only trick here is that for os.path.join to work, we have to strip the leading slash off of the URI before appending it to the webroot directory.

Aside from that, I think that I was able to clean up your cases a bit. You can investigate that if you like, but it's not as strong of a suggestion.